chore (ui) : Updated

+++ database/migrations/2024_11_22_155841_create_products_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
 
             $table->string('name');
             $table->integer('price');
-            $table->float('discount')->nullable();
+            $table->integer('discount')->nullable();
             $table->integer('stock')->nullable();
             $table->string('image')->nullable();
